+++ Performance_Shooter/src/object/Speeder.java
@@ -319,7 +319,7 @@ public class Speeder implements Entity{
 	}
 
 	@Override
-	public void damage(int amount) 
+	public void damage(float amount) 
 	{
 		
 	}
